feat. støtte BigDecimal for antallDagerPerUke og Stillingsprosent

+++ src/main/kotlin/no/nav/tiltak/tiltaknotifikasjon/avtale/AvtaleHendelseMelding.kt
@@ -2,6 +2,7 @@ package no.nav.tiltak.tiltaknotifikasjon.avtale
 
 import no.nav.tiltak.tiltaknotifikasjon.arbeidsgivernotifikasjon.NotifikasjonTekst
 import no.nav.tiltak.tiltaknotifikasjon.utils.ulid
+import java.math.BigDecimal
 import java.time.Instant
 import java.time.LocalDate
 import java.time.LocalDateTime
@@ -16,7 +17,7 @@ data class AvtaleHendelseMelding(
     val bedriftNavn: String?,
     val bedriftNr: String,
     val stillingstittel: String?,
-    val stillingprosent: Int?,
+    val stillingprosent: BigDecimal?,
     val avtaleInngått: LocalDateTime?,
     val utførtAv: String,
     val utførtAvRolle: AvtaleHendelseUtførtAvRolle,
@@ -29,7 +30,7 @@ data class AvtaleHendelseMelding(
     val sistEndret: Instant,
     val veilederNavIdent: String?,
     val annullertGrunn: String?,
-    val antallDagerPerUke: Int?,
+    val antallDagerPerUke: BigDecimal?,
     val godkjentAvDeltaker: LocalDateTime?,
     val feilregistrert: Boolean,
     val opphav: AvtaleOpphav?
